Move component into reusable directory
Also updating styles
